Moved hack to InstanceRendering module to function as a starting point
- was unable to get anything to draw on the right position, so im going to decrease the step size of the progress, it will cost more time, but has a higher chance of succeeding
